# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "contact": {
            "email": "hi.wasissubekti02@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/testimony": {
            "get": {
                "tags": [
                    "Testimony"
                ],
                "summary": "ambil semua Testimony",
                "description": "Mengambil Data Semua Testimony",
                "operationId": "testimony",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "message": "success get all testimony",
                                    "data": [
                                        {
                                            "id": 15,
                                            "image": "1801550969236851.jpg",
                                            "name": "sip ok",
                                            "business_name": "ABC",
                                            "description": "Sip",
                                            "created_at": "2024-06-10T02:36:13.000000Z",
                                            "updated_at": "2024-06-11T08:01:32.000000Z"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Testimony"
                ],
                "summary": "Create new Testimony",
                "description": "Create a new Testimony",
                "operationId": "create_testimony",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "The image of Testimony",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "The description of Testimony",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "The name of Testimony",
                                        "type": "string"
                                    },
                                    "business_name": {
                                        "description": "The business_name of Testimony",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success post Testimony",
                                    "data": {
                                        "image": "1801797620165543.jpg",
                                        "description": "lorem ipsumwfsjdkjfkjsdkfjakdjfk",
                                        "name": "sip ok",
                                        "business_name": "ABC",
                                        "created_at": "2024-06-14T01:21:56.000000Z",
                                        "updated_at": "2024-06-14T01:21:56.000000Z",
                                        "id": 21
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail post Testimony",
                                    "error": "Validation error message"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/testimony/{id}": {
            "get": {
                "tags": [
                    "Testimony"
                ],
                "summary": "Get single Testimony",
                "description": "Retrieve a single Testimony by ID",
                "operationId": "single_testimony",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Testimony",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success get Testimony",
                                    "data": {
                                        "id": 15,
                                        "image": "1801733198090700.jpg",
                                        "description": "ok",
                                        "name": "sip ok",
                                        "business_name": "ABC",
                                        "created_at": "2024-06-10T02:36:13.000000Z",
                                        "updated_at": "2024-06-13T08:17:59.000000Z"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimony not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail get Testimony",
                                    "error": "No query results for model [App\\\\Models\\\\Testimony] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Testimony"
                ],
                "summary": "Update Testimony",
                "description": "Update Testimony by ID",
                "operationId": "update_testimony",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of Testimony",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "The image of Testimony",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "description": {
                                        "description": "The description of Testimony",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "The name of Testimony",
                                        "type": "string"
                                    },
                                    "business_name": {
                                        "description": "The business_name of Testimony",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success update Testimony",
                                    "data": {
                                        "id": 1,
                                        "description": "Updated Testimony",
                                        "image": "updated_image_name.jpg",
                                        "created_at": "2024-05-18 15:52:01",
                                        "updated_at": "2024-05-18 15:52:01"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimony not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail update Testimony",
                                    "error": "No query results for model [App\\\\Models\\\\Testimony] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Testimony"
                ],
                "summary": "Delete Testimony",
                "description": "Delete Testimony by ID",
                "operationId": "delete_testimony",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the Testimony",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": true,
                                    "message": "success delete Testimony"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Testimony not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "status": false,
                                    "message": "fail delete Testimony",
                                    "error": "No query results for model [App\\\\Models\\\\Testimony] 10"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}